Pride flags: change heart emoji to heart suit character
Crucially, the heart emoji needs a color explicitly set in css
to avoid showing its graphic emoji representation, whereas the
heart suit character will just take on the color of the text because
it's not an emoji.

This is all about respecting the default client color scheme; if
it's set to white text on black background, me choosing a "shadow"
color of black won't work; but if the "shadows" just take on the
color of the text, it will be a perfect contrast color regardless.